requirements.txt
awesome-slugify==1.6.5
Flask==0.11.1
Frozen-Flask==0.12
matplotlib==1.5.1
mpld3==0.2
nose==1.3.7
numpy==1.9.2
pandas==0.18.1
